add confirmation in delete add styles

+++ app/controllers/contacts/contact/show.js
@@ -3,9 +3,11 @@ import Ember from 'ember';
 export default Ember.Controller.extend({
   actions: {
     destroy(){
-      this.get("model").destroyRecord().then(()=>{
-        this.transitionToRoute('contacts');
-      })
+      if (confirm("Want to delete?")) {
+        this.get("model").destroyRecord().then(()=>{
+          this.transitionToRoute('contacts');
+        });
+      }
     }
   }
-})
+});
